backend/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.13.0
click==8.3.2
colorama==0.4.6
fastapi==0.136.0
greenlet==3.4.0
h11==0.16.0
httptools==0.7.1
idna==3.11
numpy==2.4.4
pandas==3.0.2
psycopg==3.3.3
psycopg-binary==3.3.3
pydantic==2.13.2
pydantic_core==2.46.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.2
python-multipart==0.0.26
PyYAML==6.0.3
six==1.17.0
SQLAlchemy==2.0.49
starlette==1.0.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2026.1
uvicorn==0.44.0
watchfiles==1.1.1
websockets==16.0
